+++ packages/ngw-kit/src/index.ts
@@ -60,7 +60,13 @@ export * from './utils/utils';
 export * from './utils/identifyUtils';
 export * from './utils/getCompanyLogo';
 export * from './utils/featureLayerUtils';
+export * from './utils/fetchNgwLayerFeature';
+export * from './utils/fetchNgwLayerFeatures';
+export * from './utils/fetchNgwLayerItem';
+export * from './utils/fetchNgwLayerItems';
+export * from './utils/fetchNgwExtent';
 export * from './utils/resourceIdFromLayerOptions';
+export * from './utils/parseDateFromNgw';
 
 export * from './adapters/createBasemapWebmapItemAdapter';
 export * from './adapters/createGeoJsonAdapter';